Basic poker engine
The engine can play a match of hands, with the correct order of events
according to the poker rules. Some warning messages are written to
System.err and the hand history per hand is printed to System.out.

The communication of the engine with the bots does not include any
parameters yet.